[MetaJSON] Update to MetaJSON version 0.0.4

+++ tools/MetaJSON-Wrapper.app/Contents/Resources/AnalyserResources/cloc/Unix/t/01_opts.t
@@ -221,6 +221,27 @@ my @Tests = (
                     'ref'  => '../tests/outputs/issues/312/results.yaml',
                 },
 
+                {
+                    'name' => 'custom Smarty definition (github issue #327)',
+                    'cd'   => '../tests/inputs/issues/327',
+                    'args' => '--force-lang-def=lang.config example.smarty2',
+                    'ref'  => '../tests/outputs/issues/327/results.yaml',
+                },
+
+                {
+                    'name' => 'UTF-8 output file encoding',
+                    'cd'   => '../tests/inputs/issues/318',
+                    'args' => '--by-file --file-encoding utf8 R*.cs',
+                    'ref'  => '../tests/inputs/issues/318/Rcs.yaml',  # results in input dir
+                },
+
+                {
+                    'name' => 'distinguish TeX from VB (github issue #341)',
+                    'cd'   => '../tests/inputs/issues/341',
+                    'args' => '.',
+                    'ref'  => '../tests/outputs/issues/341/results.yaml',
+                },
+
 
 #               {
 #                   'name' => '--count-and--diff with --out',
@@ -239,20 +260,28 @@ my $Verbose = 0;
 
 my $results = 'results.yaml';
 my $work_dir = getcwd;
-my $cloc    = "$work_dir/../cloc";
+my $cloc     = "$work_dir/../cloc";   # all-purpose version
+#my $cloc     = "$work_dir/cloc";      # Unix-tuned version
 my $Run = "$cloc --quiet --yaml --out $results ";
 foreach my $t (@Tests) {
     chdir($t->{'cd'}) if defined $t->{'cd'};
+    print "Run  dir= ", cwd(), "\n" if $Verbose;
     print  $Run . $t->{'args'} if $Verbose;
     system($Run . $t->{'args'});
     ok(-e $results, $t->{'name'} . " created output");
     my %this = load_yaml($results);
-    unlink $results;
+    unlink $results unless $Verbose;
     chdir($work_dir) if defined $t->{'cd'};
+    print "Load dir= ", cwd(), "\n" if $Verbose;
     my %ref  = load_yaml($t->{'ref'});
 
 #   my $REF = LoadFile($t->{'ref'});  # using official YAML module
 #   is_deeply($REF , \%this, $t->{'name'} . " results match");
+
+#   use Data::Dumper;
+#   print Dumper(\%ref);
+#   print Dumper(\%this);
+
     is_deeply(\%ref, \%this, $t->{'name'} . " results match");
 }
 done_testing();
